Move FakePdfReportGenerator to driven/test module
Extract the fake into a shared driven/test module wired globally
via BackendKoinInitializedTest, remove redundant test dependencies
from driving/impl, and clean up empty lines in use case impl.
