+++ feat/reports/be/app/impl/build.gradle.kts
@@ -8,6 +8,7 @@ dependencies {
     implementation(project(":feat:structures:be:app:api"))
     implementation(project(":feat:findings:be:app:api"))
     implementation(project(":feat:inspections:be:app:api"))
+    testImplementation(project(":feat:reports:be:driven:test"))
     testImplementation(project(":feat:projects:be:ports"))
     testImplementation(project(":feat:structures:be:ports"))
 }
