+++ feat/reports/be/app/impl/src/test/kotlin/cz/adamec/timotej/snag/reports/be/app/impl/internal/GenerateProjectReportUseCaseImplTest.kt
@@ -19,14 +19,10 @@ import cz.adamec.timotej.snag.projects.be.model.BackendProject
 import cz.adamec.timotej.snag.projects.be.ports.ProjectsDb
 import cz.adamec.timotej.snag.projects.business.Project
 import cz.adamec.timotej.snag.reports.be.app.api.GenerateProjectReportUseCase
-import cz.adamec.timotej.snag.reports.be.ports.PdfReportGenerator
-import cz.adamec.timotej.snag.reports.be.ports.ProjectReportData
+import cz.adamec.timotej.snag.reports.be.driven.test.FakePdfReportGenerator
 import cz.adamec.timotej.snag.structures.be.ports.StructuresDb
 import cz.adamec.timotej.snag.testinfra.be.BackendKoinInitializedTest
 import kotlinx.coroutines.test.runTest
-import org.koin.core.module.dsl.singleOf
-import org.koin.dsl.bind
-import org.koin.dsl.module
 import org.koin.test.inject
 import kotlin.test.Test
 import kotlin.test.assertContentEquals
@@ -40,13 +36,6 @@ class GenerateProjectReportUseCaseImplTest : BackendKoinInitializedTest() {
     private val projectsDb: ProjectsDb by inject()
     private val structuresDb: StructuresDb by inject()
     private val fakeGenerator: FakePdfReportGenerator by inject()
-
-    override fun additionalKoinModules() =
-        listOf(
-            module {
-                singleOf(::FakePdfReportGenerator) bind PdfReportGenerator::class
-            },
-        )
 
     @Test
     fun `returns null when project does not exist`() =
@@ -132,18 +121,5 @@ class GenerateProjectReportUseCaseImplTest : BackendKoinInitializedTest() {
         private val PROJECT_ID = Uuid.parse("00000000-0000-0000-0000-000000000001")
         private val STRUCTURE_ID_1 = Uuid.parse("00000000-0000-0000-0000-000000000010")
         private val STRUCTURE_ID_2 = Uuid.parse("00000000-0000-0000-0000-000000000011")
-    }
-}
-
-internal class FakePdfReportGenerator : PdfReportGenerator {
-    var lastData: ProjectReportData? = null
-
-    override suspend fun generate(data: ProjectReportData): ByteArray {
-        lastData = data
-        return FAKE_PDF_BYTES
-    }
-
-    companion object {
-        val FAKE_PDF_BYTES = byteArrayOf(0x25, 0x50, 0x44, 0x46)
     }
 }
