+++ feat/reports/be/driven/test/build.gradle.kts
@@ -13,11 +13,3 @@
 plugins {
     alias(libs.plugins.snagBackendModule)
 }
-
-dependencies {
-    api(projects.testInfra.common)
-    implementation(projects.koinModulesAggregate.be)
-    implementation(projects.feat.shared.database.be.test)
-    implementation(projects.lib.storage.be.test)
-    implementation(projects.feat.reports.be.driven.test)
-}
