+++ feat/reports/be/driven/test/src/main/kotlin/cz/adamec/timotej/snag/reports/be/driven/test/ReportsDrivenTestModule.kt
@@ -1,0 +1,23 @@
+/*
+ * Copyright (c) 2026 Timotej Adamec
+ * SPDX-License-Identifier: MIT
+ *
+ * This file is part of the thesis:
+ * "Multiplatform snagging system with code sharing maximisation"
+ *
+ * Czech Technical University in Prague
+ * Faculty of Information Technology
+ * Department of Software Engineering
+ */
+
+package cz.adamec.timotej.snag.reports.be.driven.test
+
+import cz.adamec.timotej.snag.reports.be.ports.PdfReportGenerator
+import org.koin.core.module.dsl.singleOf
+import org.koin.dsl.bind
+import org.koin.dsl.module
+
+val reportsDrivenTestModule =
+    module {
+        singleOf(::FakePdfReportGenerator) bind PdfReportGenerator::class
+    }
