+++ settings.gradle.kts
@@ -178,6 +178,7 @@ include(":feat:reports:be:app:api")
 include(":feat:reports:be:app:impl")
 include(":feat:reports:be:ports")
 include(":feat:reports:be:driven:impl")
+include(":feat:reports:be:driven:test")
 include(":feat:reports:be:driving:impl")
 include(":feat:shared:storage:be")
 include(":feat:shared:storage:fe")
