+++ testInfra/be/src/main/kotlin/cz/adamec/timotej/snag/testinfra/be/BackendKoinInitializedTest.kt
@@ -15,6 +15,7 @@ package cz.adamec.timotej.snag.testinfra.be
 import cz.adamec.timotej.snag.di.aggregate.be.backendModulesAggregate
 import cz.adamec.timotej.snag.feat.shared.database.be.di.sharedDatabaseTestModule
 import cz.adamec.timotej.snag.lib.storage.be.test.di.fakeStorageTestModule
+import cz.adamec.timotej.snag.reports.be.driven.test.reportsDrivenTestModule
 import cz.adamec.timotej.snag.testinfra.KoinInitializedTest
 import org.koin.core.module.Module
 
@@ -24,5 +25,6 @@ abstract class BackendKoinInitializedTest : KoinInitializedTest() {
             backendModulesAggregate,
             sharedDatabaseTestModule,
             fakeStorageTestModule,
+            reportsDrivenTestModule,
         )
 }
